refactor: folder structure of th ehandlers
