+++ src/labs/acts/__init__.py
@@ -1,2 +1,2 @@
-"""
+"""Aggregated acts
 """
