+++ src/labs/acts/ext/scenes/__init__.py
@@ -1,0 +1,15 @@
+"""
+"""
+
+from fastapi import APIRouter, Request, status
+
+router = APIRouter(tags=["ext"])
+
+@router.get("/echo")
+async def echo(request: Request):
+    """Echo back a response to say hello.
+
+    Purpose of this endpoint is to echo back what was received, this merely
+    validated that the server is up and running.
+    """
+    return {"hello": "world"}
